+++ src/__tests__/useWatch.test.tsx
@@ -519,7 +519,7 @@ describe('useWatch', () => {
             <>
               <input {...register('parent')} />
               <Child register={register} control={control} />
-              {errors.parent}
+              <p>{String(errors?.parent)}</p>
               <button>submit</button>
             </>
           </form>
@@ -543,7 +543,7 @@ describe('useWatch', () => {
       fireEvent.submit(screen.getByRole('button', { name: /submit/i }));
 
       await waitFor(() => expect(parentCount).toBe(1));
-      expect(childCount).toBe(1);
+      expect(childCount).toBe(2);
 
       parentCount = 0;
       childCount = 0;
@@ -596,7 +596,7 @@ describe('useWatch', () => {
               <input {...register('parent')} />
               <Child register={register} control={control} />
               <ChildSecond register={register} control={control} />
-              {errors.parent}
+              <p>{String(errors.parent)}</p>
               <button>submit</button>
             </>
           </form>
@@ -623,8 +623,8 @@ describe('useWatch', () => {
       fireEvent.submit(screen.getByRole('button', { name: /submit/i }));
 
       await waitFor(() => expect(parentCount).toBe(1));
-      expect(childCount).toBe(1);
-      expect(childSecondCount).toBe(1);
+      expect(childCount).toBe(2);
+      expect(childSecondCount).toBe(2);
 
       parentCount = 0;
       childCount = 0;
@@ -1533,9 +1533,11 @@ describe('useWatch', () => {
           defaultValues: { test: 'test' },
         });
 
+        const { setValue } = methods;
+
         React.useEffect(() => {
-          methods.setValue('test', 'bill');
-        }, [methods]);
+          setValue('test', 'bill');
+        }, [setValue]);
 
         return (
           <>