+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [7.54.0] - 2024-12-09
+
+### Changed
+
+- `useForm` return methods will be memorized based `formState` update
+
 ## [7.53.0] - 2024-8-31
 
 ### Added
@@ -17,7 +23,9 @@
 - added 'validateFields' to formState
 
 ```tsx
-const { formState: { validateFields } } = useForm();
+const {
+  formState: { validateFields },
+} = useForm();
 ```
 
 ## [7.49.0] - 2023-12-10
@@ -342,7 +350,8 @@ function App() {
   return (
     <form>
       <CustomFormComponent control={control} name="foo" /> {/* no error */}
-      <CustomFormComponent control={control} name="baz" /> {/*  throw an error since baz is string */}
+      <CustomFormComponent control={control} name="baz" />{' '}
+      {/*  throw an error since baz is string */}
     </form>
   );
 }
@@ -825,7 +834,8 @@ const App = () => {
           },
         })}
       />
-      <input {...register('lastName', { deps: ['firstName'] })} /> // dependant validation
+      <input {...register('lastName', { deps: ['firstName'] })} /> // dependant
+      validation
     </form>
   );
 };